tinydm-run-session: back up previous log file so it's not overwritten (MR 7)
Sometimes the DE crashes, causing tinydm to immediately restart. When
this happens, any useful info in the log file is overwritten and lost
forever. This change causes tinydm to save the log file from the
previous session. It only saves N-1 logs.

This behavior is similar to what X11 does with Xorg.0.log and
Xorg.0.log.old.